pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "animal_id"
version = "0.1.0"
description = "Animal identification and tracking pipeline"
authors = [
  { name = "Ryan Peterson" },
]
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.0.0",
    "opencv-python>=4.8.0",
    "onnxruntime>=1.15.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "Pillow>=10.0.0",
    "transformers>=4.30.0",
    "requests>=2.31.0",
    "pathlib2>=2.3.7",
    "wandb>=0.23.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "jupyter>=1.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["animal_id*"]

[tool.ruff]
target-version = "py312"
exclude = ["immich-clone", ".gemini"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]


